Hero can collide with enemies
Hero can collide with enemies by raycosting. Setting ray-cast parameter to 1 means it sees everything on layer 0 and
will therefore avoid a collision. If two bodies overlap and at least one
of their mask's recognizes the other's layer, a collision occurs. Next,
the collision must be recognized in the code.