Follower AI improvements
Follower determines hero positions before finding best path regardless
of hero movement. Previously, the hero could change positions while the
follower was trying to determine the best path.

Follower will find best path to within one tile of hero instead of hero
exactly.  Without this, if one hero was between two tiles and the other
was on a tile exactly, the follower would go for the latter regardless
of steps because the distance was zero.